requirements.txt
numpy
pillow
pyyaml
requests
torch
torchvision
tweepy
selenium